+++ functions/Get-PSBlueSkyFollowing.ps1
@@ -1,36 +1,67 @@
 #https://docs.bsky.app/docs/api/app-bsky-graph-get-follows
 Function Get-BskyFollowing {
-    [CmdletBinding()]
+    [CmdletBinding(DefaultParameterSetName = 'Limit')]
     [OutputType('PSBlueskyFollowProfile')]
     Param(
-        [Parameter(HelpMessage = "Enter the number of accounts that you follow to retrieve between 1 and 100. Default is 50.")]
-        [ValidateRange(1,100)]
+        [Parameter(
+            Position = 0,
+            HelpMessage = 'Enter the number of accounts that you follow to retrieve between 1 and 100. Default is 50.',
+            ParameterSetName = 'Limit'
+        )]
+        [ValidateRange(1, 100)]
         [int]$Limit = 50,
+        [Parameter(ParameterSetName = 'All', HelpMessage = 'Return All followers')]
+        [switch]$All,
         [Parameter(Mandatory, HelpMessage = 'A PSCredential with your Bluesky username and password')]
         [PSCredential]$Credential
     )
 
     Begin {
         Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Starting $($MyInvocation.MyCommand)"
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Using PowerShell version $($PSVersionTable.PSVersion)"
+        if ($MyInvocation.CommandOrigin -eq 'Runspace') {
+            #Hide this metadata when the command is called from another command
+            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running module version $ModuleVersion"
+            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Using PowerShell version $($PSVersionTable.PSVersion)"
+            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running on $($PSVersionTable.OS)"
+        }
         $token = Get-BskyAccessToken -Credential $Credential
         $UserName = $Credential.UserName
     } #begin
     Process {
         If ($token) {
-            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Querying $limit following accounts for $Username"
+            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Querying $limit response accounts for $Username"
 
-            $apiUrl = "$PDSHOST/xrpc/app.bsky.graph.getFollows?limit=$Limit&actor=$UserName"
+            if ($PSCmdlet.ParameterSetName -eq 'Limit') {
+                $apiUrl = "$PDSHOST/xrpc/app.bsky.graph.getFollows?limit=$Limit&actor=$UserName"
+            }
+            else {
+                $apiUrl = "$PDSHOST/xrpc/app.bsky.graph.getFollows?limit=100&actor=$UserName"
+            }
             Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Processing: $apiUrl"
             $headers = @{
                 Authorization  = "Bearer $token"
                 'Content-Type' = 'application/json'
             }
 
-            $Following = Invoke-RestMethod -Uri $apiUrl -Method Get -Headers $headers
-            If ($Following) {
-                Write-Information -MessageData $Following -Tags raw
-                foreach ($profile in $Following.Follows) {
+            $results = @()
+            $response = Invoke-RestMethod -Uri $apiUrl -Method Get -Headers $headers
+            If ($response) {
+                $results += $response.follows
+                Write-Information -MessageData $response -Tags raw
+                if ($PSCmdlet.ParameterSetName -eq 'All') {
+                    Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS]  Getting all following accounts from $($response.cursor)"
+                    # iterate remaining pages using 'cursor' response value
+                    while ($response.cursor) {
+                        $url = $apiUrl + "&cursor=$($response.cursor)"
+                        $response = Invoke-RestMethod -Uri $url -Method Get -Headers $headers
+                        If ($response.follows) {
+                            $results += $response.follows
+                        }
+                    }
+                } #If All
+
+                 Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Processing $($results.Count) accounts"
+                foreach ($profile in $results) {
                     [PSCustomObject]@{
                         PSTypeName  = 'PSBlueskyFollowProfile'
                         Username    = $profile.handle